+++ packages/rpc-core/package.json
@@ -67,6 +67,7 @@
         "@solana/codecs-strings": "workspace:*",
         "@solana/eslint-config-solana": "^1.0.2",
         "@solana/keys": "workspace:*",
+        "@solana/rpc-json-parsed": "workspace:*",
         "@solana/rpc-transport": "workspace:*",
         "@solana/rpc-types": "workspace:*",
         "@solana/transactions": "workspace:*",
