+++ packages/rpc-json-parsed/package.json
@@ -60,7 +60,9 @@
         "maintained node versions"
     ],
     "devDependencies": {
+        "@solana/addresses": "workspace:*",
         "@solana/eslint-config-solana": "^1.0.2",
+        "@solana/rpc-types": "workspace:*",
         "@swc/jest": "^0.2.29",
         "@types/jest": "^29.5.11",
         "@typescript-eslint/eslint-plugin": "^6.13.2",
