+++ packages/rpc-json-parsed/src/index.ts
@@ -1,0 +1,1 @@
+export * from './token-accounts';
