+++ packages/rpc-types/src/index.ts
@@ -2,4 +2,5 @@ export * from './commitment';
 export * from './lamports';
 export * from './stringified-bigint';
 export * from './stringified-number';
+export * from './token-amount';
 export * from './unix-timestamp';
