+++ packages/rpc-types/src/token-amount.ts
@@ -1,0 +1,9 @@
+import { StringifiedBigInt } from './stringified-bigint';
+import { StringifiedNumber } from './stringified-number';
+
+export type TokenAmount = Readonly<{
+    amount: StringifiedBigInt;
+    decimals: number;
+    uiAmount: number | null;
+    uiAmountString: StringifiedNumber;
+}>;
